// src/interfaces/Actions/ConditionalStatement.tsx
// Generated by https://quicktype.io

export interface ConditionalStatement {
    Type:               string;
    ExpressionList:     ExpressionList[];
    ActionSetOnTrue:    ActionSetOn;
    ActionSetOnFalse:   ActionSetOn;
    Notes:              string;
    ActionId:           number;
    Disabled:           boolean;
    InValid:            boolean;
    ValidationMessages: any[];
    Metadata:           null;
}

export interface ActionSetOn {
    Name:         string;
    ActionSetId:  string;
    Key:          string;
    Scope:        number;
    Actions:      any[];
    Notes:        null;
    IsInAppStore: boolean;
}

export interface ExpressionList {
    Operation:         number | null;
    DateField:         null;
    Source:            Source | null;
    AdditionalSources: null;
    DateWhere:         null;
}

export interface Source {
    ValueType:         number;
    Value:             string;
    Property:          string;
    ChildProperty:     null;
    ValidationMessage: null;
    FontStyles:        null;
}